requirements.txt
torch
numpy
Pillow
pygetwindow
mss